requirements.txt
numba==0.55.1
intel-numpy==1.21.4
pandas==1.5.1
scipy==1.9.3
h5py==3.7.0
scikit-learn==1.1.3
joblib==1.2.0
librosa==0.9.2
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.0+cu116
torchaudio==0.13.0
torchmetrics==0.8.1
pytorch_lightning==1.8.4
implicit==0.6.1
tqdm==4.64.1
soundfile==0.11.0
hdpgmm @ git+https://github.com/eldrin/pytorch-hdpgmm@ismir22
musmtl @ git+https://github.com/eldrin/MTLMusicRepresentation-PyTorch@master
# clmr @ git+https://github.com/eldrin/CLMR@hdpgmm_study_mod
